Introduce `LowLevel` base class, refactor low-level utilities to enforce inheritance and alignment with framework standards
- **Added `LowLevel` base class**:
  - Introduced an abstract `LowLevel` class in `FireHub\Core\Support` to act as the foundational base class for all low-level helpers.
  - Enforced stateless, static-only usage by utilizing a `final protected` constructor to prevent instantiation and subclassing outside the intended use case.
  - Marked the class as `@internal` to restrict usage to FireHub's internal ecosystem.

- **Refactored low-level utility classes** to extend the `LowLevel` base class:
  - Updated the following classes to inherit from `LowLevel`: `CharMB`, `CharSB`, `StrMB`, `Data`, `DataIs`, `Math`, `ClsObj`, and `Arr`.
  - Added `use` statements to reference the `LowLevel` base class for consistency.
  - Documented each class with a `@note` annotation, explicitly warning against instantiation or extension outside FireHub's ecosystem.

- **Standardized docblock annotations**:
  - Replaced hardcoded Git commit checksums in the `@version` tags with standardized `$Id$` placeholders across all refactored classes.
  - Improved clarity and consistency in class-level and method-level docblocks by fixing phrasing and ensuring adherence to professional documentation standards.
  - Explicitly marked all helper classes as `@internal`.

- **Updated `NoNativeFunctionsRule` in PHPStan**:
  - Refactored the rule to leverage the `LowLevel` inheritance hierarchy for internal function validations, simplifying function-exclusion logic with a class-based reflection check.

These changes introduce a centralized, enforced design standard for FireHub’s low-level utilities by unifying them under a cohesive base class. The refactoring ensures improved maintainability, consistency, and adherence to internal framework guidelines.
