+++ src/support/lowlevel/firehub.CharMB.php
@@ -15,6 +15,7 @@
 
 namespace FireHub\Core\Support\LowLevel;
 
+use FireHub\Core\Support\LowLevel;
 use FireHub\Core\Shared\Enums\String\Encoding;
 use FireHub\Core\Throwable\Error\LowLevel\String\InvalidEncodingError;
 
@@ -28,8 +29,11 @@ use function mb_ord;
  * @since 1.0.0
  *
  * @internal
+ *
+ * @note This class is intended only as an inheritance base for framework-internal helpers.<br>
+ * Do not instantiate or extend outside the FireHub low-level helper ecosystem.
  */
-final class CharMB {
+final class CharMB extends LowLevel {
 
     /**
      * ### Return character by Unicode code point value
