+++ src/support/lowlevel/firehub.CharSB.php
@@ -15,6 +15,7 @@
 
 namespace FireHub\Core\Support\LowLevel;
 
+use FireHub\Core\Support\LowLevel;
 use FireHub\Core\Throwable\Error\LowLevel\String\CodepointOutsideValidRangeError;
 
 use function chr;
@@ -27,8 +28,11 @@ use function ord;
  * @since 1.0.0
  *
  * @internal
+ *
+ * @note This class is intended only as an inheritance base for framework-internal helpers.<br>
+ * Do not instantiate or extend outside the FireHub low-level helper ecosystem.
  */
-final class CharSB {
+final class CharSB extends LowLevel {
 
     /**
      * ### Generate a single-byte string from a number
